- Update Hot Update

+++ Assets/Scripts/Addressables/AddressablesUpdater.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine.ResourceManagement.ResourceProviders;
 using UnityEngine.ResourceManagement.ResourceLocations;
+using UnityEngine.AddressableAssets.ResourceLocators;
 using UnityEngine.ResourceManagement.Util;
 using UnityEngine.AddressableAssets;
 using System.Collections.Generic;
@@ -8,7 +9,7 @@ using UnityEngine;
 using System.IO;
 using System;
 
-public enum RequestDownloadFeedback
+public enum RequestDownloadResult
 {
     None,
     Agree,
@@ -17,16 +18,22 @@ public enum RequestDownloadFeedback
 
 public class AddressablesUpdater : MonoBehaviour
 {
+    public static RequestDownloadResult Result;
     public static event Action OnCompleted;
     public static event Action<long, long> OnDownload;
     public static event Action<long> OnDownloadCompleted;
     public static Func<IEnumerator> AfterDownloadHandle;
-    public static Func<RequestDownloadFeedback> RequestDownloadHandle;
+    public static Func<long, IEnumerator> RequestDownloadHandle;
+
+    private long downloadSize = 0;
+    private IList<IResourceLocator> updateLocators;
+    private Dictionary<int, long> sizeMap = new Dictionary<int, long>();
+    private Dictionary<string, long> bundleSizeMap = new Dictionary<string, long>();
 
     private static string addressablesDir = Application.persistentDataPath + "/" + addressablesFolder;
-    private static string backupDir = Application.persistentDataPath + "/" + backupFolder;
+    private static string originalDir = Application.persistentDataPath + "/" + originalFolder;
     private const string addressablesFolder = "com.unity.addressables";
-    private const string backupFolder = "backup.addressables";
+    private const string originalFolder = "original.addressables";
 
     private IEnumerator Start()
     {
@@ -45,89 +52,145 @@ public class AddressablesUpdater : MonoBehaviour
         var catalogs = checkHandle.Result;
         Addressables.Release(checkHandle);
         if (catalogs == null || catalogs.Count <= 0) { yield break; }
-        SaveCatalogs();
-        var updateHandle = Addressables.UpdateCatalogs(catalogs, false);
-        yield return updateHandle;
-        long downloadSize = 0;
-        long size = 0;
-        var sizeMap = new Dictionary<int, long>();
-        var locators = updateHandle.Result;
-        Addressables.Release(updateHandle);
-        for (int i = 0; i < locators.Count; i++)
+        SaveOriginalCatalogs();
+        yield return StartCoroutine(InitializeBundleSizeMapAsync());
+        yield return StartCoroutine(UpdateCatalogsAsync(catalogs));
+        CalculateUpdateSize();
+        if (downloadSize <= 0) { yield break; }
+        yield return StartCoroutine(RequestDownloadHandle?.Invoke(downloadSize));
+        if (Result == RequestDownloadResult.Disagree)
         {
+            yield return StartCoroutine(CancelDownloadAsync());
+            yield break;
+        }
+        DeleteOriginalCatalogs();
+        yield return StartCoroutine(DownloadAsync());
+        OnDownloadCompleted?.Invoke(downloadSize);
+        yield return StartCoroutine(AfterDownloadHandle?.Invoke());
+    }
+
+    private void OnDestroy()
+    {
+        LoadOriginalCatalogs();
+    }
+
+    private IEnumerator InitializeBundleSizeMapAsync()
+    {
+        bundleSizeMap.Clear();
+        if (!Directory.Exists(addressablesDir)) { yield break; }
+        var catalogPaths = GetCatalogs(addressablesDir, "json", SearchOption.TopDirectoryOnly);
+        for (int i = 0; i < catalogPaths.Length; i++)
+        {
+            var catalogHandle = Addressables.LoadContentCatalogAsync(catalogPaths[i]);
+            yield return catalogHandle;
             IList<IResourceLocation> locations;
-            var e = locators[i].Keys.GetEnumerator();
-            var j = 0;
+            var locator = catalogHandle.Result;
+            var e = locator.Keys.GetEnumerator();
             while (e.MoveNext())
             {
-                if (locators[i].Locate(e.Current, typeof(object), out locations))
+                if (locator.Locate(e.Current, typeof(object), out locations))
                 {
                     for (int k = 0; k < locations.Count; k++)
                     {
-                        if (locations[k].Data != null && locations[k].Data is AssetBundleRequestOptions && ResourceManagerConfig.IsPathRemote(locations[k].InternalId))
+                        if (bundleSizeMap.ContainsKey(locations[k].InternalId)) { continue; }
+                        if (ResourceManagerConfig.IsPathRemote(locations[k].InternalId) && locations[k].Data != null && locations[k].Data is AssetBundleRequestOptions)
                         {
-                            size += (locations[k].Data as AssetBundleRequestOptions).BundleSize;
+                            bundleSizeMap.Add(locations[k].InternalId, 0);
                         }
                     }
-                    sizeMap.Add(j, size);
-                    downloadSize += size;
-                    size = 0;
                 }
-                j++;
             }
+            Addressables.Release(catalogHandle);
         }
-        if (downloadSize <= 0) { yield break; }
-        while (RequestDownloadHandle?.Invoke() == RequestDownloadFeedback.None)
-        {
-            yield return null;
-        }
-        if (RequestDownloadHandle?.Invoke() == RequestDownloadFeedback.Disagree)
-        {
-            LoadCatalogs();
-            for (int i = 0; i < locators.Count; i++)
-            {
-                Addressables.RemoveResourceLocator(locators[i]);
-            }
-            var catalogPaths = GetCatalogs(addressablesDir, "json", SearchOption.TopDirectoryOnly);
-            for (int i = 0; i < catalogPaths.Length; i++)
-            {
-                var catalogHandle = Addressables.LoadContentCatalogAsync(catalogPaths[i]);
-                yield return catalogHandle;
-                Addressables.AddResourceLocator(catalogHandle.Result);
-                Addressables.Release(catalogHandle);
-            }
-#if UNITY_EDITOR
-            Debug.Log(string.Format("Cancel Download Size: {0}", downloadSize));
-#endif
-            yield break;
-        }
-        DeleteCatalogs();
-        for (int i = 0; i < locators.Count; i++)
+    }
+
+    private IEnumerator UpdateCatalogsAsync(List<string> catalogs)
+    {
+        var updateHandle = Addressables.UpdateCatalogs(catalogs, false);
+        yield return updateHandle;
+        downloadSize = 0;
+        sizeMap = new Dictionary<int, long>();
+        updateLocators = updateHandle.Result;
+        Addressables.Release(updateHandle);
+    }
+
+    private void CalculateUpdateSize()
+    {
+        for (int i = 0; i < updateLocators.Count; i++)
         {
             IList<IResourceLocation> locations;
-            var e = locators[i].Keys.GetEnumerator();
-            var j = 0;
+            var e = updateLocators[i].Keys.GetEnumerator();
+            var index = 0;
             while (e.MoveNext())
             {
-                if (locators[i].Locate(e.Current, typeof(object), out locations))
+                long size = 0;
+                if (updateLocators[i].Locate(e.Current, typeof(object), out locations))
+                {
+                    for (int k = 0; k < locations.Count; k++)
+                    {
+                        if (bundleSizeMap.ContainsKey(locations[k].InternalId)) { continue; }
+                        if (ResourceManagerConfig.IsPathRemote(locations[k].InternalId) && locations[k].Data != null && locations[k].Data is AssetBundleRequestOptions)
+                        {
+                            long bundleSize = (locations[k].Data as AssetBundleRequestOptions).BundleSize;
+                            size += bundleSize;
+                            bundleSizeMap.Add(locations[k].InternalId, bundleSize);
+                        }
+                    }
+                }
+                sizeMap.Add(index, size);
+                downloadSize += size;
+                index++;
+            }
+        }
+    }
+
+    private IEnumerator CancelDownloadAsync()
+    {
+        LoadOriginalCatalogs();
+        for (int i = 0; i < updateLocators.Count; i++)
+        {
+            Addressables.RemoveResourceLocator(updateLocators[i]);
+        }
+        var catalogPaths = GetCatalogs(addressablesDir, "json", SearchOption.TopDirectoryOnly);
+        for (int i = 0; i < catalogPaths.Length; i++)
+        {
+            var catalogHandle = Addressables.LoadContentCatalogAsync(catalogPaths[i]);
+            yield return catalogHandle;
+            Addressables.AddResourceLocator(catalogHandle.Result);
+            Addressables.Release(catalogHandle);
+        }
+#if UNITY_EDITOR
+        Debug.Log(string.Format("Cancel Download Size: {0}", downloadSize));
+#endif
+    }
+
+    private IEnumerator DownloadAsync()
+    {
+        long size = 0;
+        for (int i = 0; i < updateLocators.Count; i++)
+        {
+            IList<IResourceLocation> locations;
+            var e = updateLocators[i].Keys.GetEnumerator();
+            var index = 0;
+            while (e.MoveNext())
+            {
+                if (updateLocators[i].Locate(e.Current, typeof(object), out locations))
                 {
                     var downloadHandle = Addressables.DownloadDependenciesAsync(locations);
                     while (downloadHandle.PercentComplete < 1 && !downloadHandle.IsDone)
                     {
-                        OnDownload?.Invoke(size + (long)(downloadHandle.PercentComplete * sizeMap[j]), downloadSize);
+                        OnDownload?.Invoke(size + (long)(downloadHandle.PercentComplete * sizeMap[index]), downloadSize);
 #if UNITY_EDITOR
-                        Debug.Log(string.Format("Download Size: {0}/{1}", size + (long)(downloadHandle.PercentComplete * sizeMap[j]), downloadSize));
+                        Debug.Log(string.Format("Download Size: {0}/{1}", size + (long)(downloadHandle.PercentComplete * sizeMap[index]), downloadSize));
 #endif
                         yield return null;
                     }
-                    size += sizeMap[j];
+                    size += sizeMap[index];
                     Addressables.Release(downloadHandle);
                 }
-                j++;
+                index++;
             }
         }
-        OnDownloadCompleted?.Invoke(downloadSize);
-        yield return StartCoroutine(AfterDownloadHandle?.Invoke());
     }
 
     private string[] GetCatalogs(string path, string suffix, SearchOption searchOption)
@@ -151,40 +214,38 @@ public class AddressablesUpdater : MonoBehaviour
         return result;
     }
 
-    private void SaveCatalogs()
+    private void SaveOriginalCatalogs()
     {
         if (!Directory.Exists(addressablesDir)) { return; }
         var catalogs = GetCatalogs(addressablesDir, SearchOption.TopDirectoryOnly);
-        if (!Directory.Exists(backupDir)) { Directory.CreateDirectory(backupDir); }
+        if (!Directory.Exists(originalDir)) { Directory.CreateDirectory(originalDir); }
         for (int i = 0; i < catalogs.Length; i++)
         {
-            File.Copy(catalogs[i], catalogs[i].Replace(addressablesFolder, backupFolder));
+            File.Copy(catalogs[i], catalogs[i].Replace(addressablesFolder, originalFolder));
         }
     }
 
-    private void LoadCatalogs()
+    private void LoadOriginalCatalogs()
     {
-        if (!Directory.Exists(addressablesDir)) { return; }
+        if (!Directory.Exists(addressablesDir) || !Directory.Exists(originalDir)) { return; }
         var catalogs = GetCatalogs(addressablesDir, SearchOption.TopDirectoryOnly);
         for (int i = 0; i < catalogs.Length; i++)
         {
             File.Delete(catalogs[i]);
         }
-        if (!Directory.Exists(backupDir)) { return; }
-        catalogs = GetCatalogs(backupDir, SearchOption.TopDirectoryOnly);
+        catalogs = GetCatalogs(originalDir, SearchOption.TopDirectoryOnly);
         for (int i = 0; i < catalogs.Length; i++)
         {
-            File.Move(catalogs[i], catalogs[i].Replace(backupFolder, addressablesFolder));
+            File.Move(catalogs[i], catalogs[i].Replace(originalFolder, addressablesFolder));
         }
+        DeleteOriginalCatalogs();
     }
 
-    private void DeleteCatalogs()
+    private void DeleteOriginalCatalogs()
     {
-        if (!Directory.Exists(backupDir)) { return; }
-        var catalogs = GetCatalogs(backupDir, SearchOption.TopDirectoryOnly);
-        for (int i = 0; i < catalogs.Length; i++)
+        if (Directory.Exists(originalDir))
         {
-            File.Delete(catalogs[i]);
+            Directory.Delete(originalDir, true);
         }
     }
 }
